fix: Q1_0_g128 CPU dot product int truncation
The accumulator `sumi` in ggml_vec_dot_q1_0_g128_q8_0 was declared
as `int` but accumulates `float d1 * int sumi_block`, causing the
float result to be truncated to integer on each iteration. This
produced garbage output for Q1_0_g128 models on CPU.

Fix: change `int sumi = 0` to `float sumi = 0` in both the x86
and generic (portable) kernels.

+++ ggml/src/ggml-cpu/arch/x86/quants.c
@@ -670,7 +670,7 @@ void ggml_vec_dot_q1_0_g128_q8_0(int n, float * GGML_RESTRICT s, size_t bs, cons
     for (int ib = 0; ib < nb; ++ib) {
         const float d0 = GGML_CPU_FP16_TO_FP32(x[ib].d);
         
-        int sumi = 0;
+        float sumi = 0;
         
         // Process 4 Q8_0 blocks (4 * 32 = 128 elements)
         for (int k = 0; k < 4; k++) {
